Add sanitizers, expand server and user functionalities
Sanitizers are added to CMakeLists.txt to prevent memory leaks and undefined behaviors. Expanded server and user functionalities in the code to create a more robust application. These changes include adding a socket exception in the server and handling request methods in both server and user class. The main.cpp is also modified to instantiate and run the server.